Reorganize modules
Move webdriver-related code from `helpers` and `chromeSpoofing` to `webdriver-wrapper`.
Remove global variables.
Rename functions and wrap cetrain function calls to better readability.
Document internal functions of new `webdriver-wrapper` module.
