+++ tests/components/connector-spec.js
@@ -1,25 +1,25 @@
 'use strict';
 
-/* global helpers */
-
 const DEFAULT_RECOGNIZE_TIMEOUT = 30000;
 const WAIT_FOR_PLAYER_ELEMENT_TIMEOUT = 10000;
+
+const helpers = require('./../helpers/helpers');
 const webdriver = require('selenium-webdriver');
 
 /**
  * Test if website can be loaded successfully
- * @param  {Object} driver Webdriver instance
+ * @param  {Object} driver WebDriverWrapper instance
  * @param  {Array} options Options (see below)
  *
  * Options
  * @param  {String} url Website URL
  * @param  {Number} urlLoadTimeout URL load timeout is milliseconds
  */
-module.exports.loadSite = function(driver, options) {
+module.exports.shouldLoadWebsite = function(driver, options) {
 	var opts = options || {};
 
 	it('should load website', function(done) {
-		helpers.getAndWait(driver, opts.url, opts.urlLoadTimeout).then(function() {
+		driver.load(opts.url, opts.urlLoadTimeout).then(function() {
 			done();
 		}, function(err) {
 			done(err);
@@ -29,18 +29,18 @@ module.exports.loadSite = function(driver, options) {
 
 /**
  * Test if play button can be clicked
- * @param  {Object} driver Webdriver instance
+ * @param  {Object} driver WebDriverWrapper instance
  * @param  {Array} options Options (see below)
  *
  * Options
  * @param  {String} playButtonSelector CSS selector of play button
  */
-module.exports.clickPlayButton = function(driver, options) {
+module.exports.shouldClickPlayButton = function(driver, options) {
 	var opts = options || {};
 
 	if (opts.playButtonSelector) {
 		it('should play a song', function(done) {
-			helpers.promiseClick(driver, {css: opts.playButtonSelector}).then(function() {
+			driver.click({css: opts.playButtonSelector}).then(function() {
 				done();
 			}, function(err) {
 				done(err);
@@ -52,13 +52,13 @@ module.exports.clickPlayButton = function(driver, options) {
 
 /**
  * Test if a now playing song is correctly recognized
- * @param  {Object} driver Webdriver instance
+ * @param  {Object} driver WebDriverWrapper instance
  * @param  {Array} options Options (see below)
  *
  * Options
  * @param  {Number} recognizeTimeout Recognize timeout in milliseconds
  */
-module.exports.recognizeSong = function(driver, options) {
+module.exports.shouldRecogniseSong = function(driver, options) {
 	it('should recognise a playing song', function(done) {
 		promiseRecognizeSong(driver, options).then(function() {
 			done();
@@ -70,15 +70,15 @@ module.exports.recognizeSong = function(driver, options) {
 
 /**
  * Load website and check if player element exists.
- * @param  {Object} driver Webdriver instance
+ * @param  {Object} driver WebDriverWrapper instance
  * @param  {Array} options Options (see below)
  *
  * Options
- * @see {@link loadSite}
+ * @see {@link shouldLoadWebsite}
  */
-exports.loadCheckPlayer = function(driver, options) {
+exports.shouldContainPlayerElement = function(driver, options) {
 	it('should load website and check player element', function(done) {
-		promiseLoadCheckPlayer(driver, options).then(function() {
+		promiseCheckPlayerElement(driver, options).then(function() {
 			done();
 		}, function(err) {
 			done(err);
@@ -89,18 +89,17 @@ exports.loadCheckPlayer = function(driver, options) {
 /**
  * Perform a complex test for website. Includes load test,
  * play button click test and song recongnition test.
- * @param  {Object} driver Webdriver instance
+ * @param  {Object} driver WebDriverWrapper instance
  * @param  {Array} options Options (see below)
  *
  * Options
- * @param  {String} url Website URL
- * @param  {Number} urlLoadTimeout URL load timeout is milliseconds
- * @param  {String} playButtonSelector CSS selector of play button
- * @param  {Number} recognizeTimeout Recognize timeout in milliseconds
+ * @see {@link shouldLoadWebsite}
+ * @see {@link shouldClickPlayButton}
+ * @see {@link shouldRecogniseSong}
  */
-module.exports.loadPlayListen = function(driver, options) {
+module.exports.shouldBehaveLikeMusicSite = function(driver, options) {
 	it('should load site and recognize a song', function(done) {
-		promiseLoadPlayListen(driver, options).then(function() {
+		promiseBehaveLikeMusicSite(driver, options).then(function() {
 			done();
 		}, function(err) {
 			done(err);
@@ -110,25 +109,37 @@ module.exports.loadPlayListen = function(driver, options) {
 
 /* Internal */
 
+/**
+ * @see {@link shouldLoadWebsite}
+ * @return {Promise} Promise that will be resolved when the task has completed
+ */
 function promiseLoadSite(driver, options) {
 	var opts = options || {};
-	return helpers.getAndWait(driver, opts.url);
+	return driver.load(opts.url);
 }
 
+/**
+ * @see {@link shouldClickPlayButton}
+ * @return {Promise} Promise that will be resolved when the task has completed
+ */
 function promiseClickPlayButton(driver, options) {
 	var opts = options || {};
 
 	if (opts.playButtonSelector) {
-		return helpers.promiseClick(driver, {css: opts.playButtonSelector});
+		return driver.click({css: opts.playButtonSelector});
 	} else {
 		return webdriver.promise.fulfilled();
 	}
 }
 
+/**
+ * @see {@link shouldRecogniseSong}
+ * @return {Promise} Promise that will be resolved when the task has completed
+ */
 function promiseRecognizeSong(driver, options) {
 	var opts = options || {};
 	var timeout = opts.recognizeTimeout || DEFAULT_RECOGNIZE_TIMEOUT;
-	return helpers.listenFor(driver, 'connector_state_changed', timeout).then(function(res) {
+	return driver.waitForSongRecognition(timeout).then(function(res) {
 		var song = res.data;
 
 		if (global.DEBUG) {
@@ -143,20 +154,28 @@ function promiseRecognizeSong(driver, options) {
 	});
 }
 
-function promiseLoadCheckPlayer(driver, options) {
+/**
+ * @see {@link shouldContainPlayerElement}
+ * @return {Promise} Promise that will be resolved when the task has completed
+ */
+function promiseCheckPlayerElement(driver, options) {
 	var opts = options || {};
 
 	return promiseLoadSite(driver, opts).then(function() {
 		return promiseClickPlayButton(driver, opts);
 	}).then(function() {
 		var timeout = WAIT_FOR_PLAYER_ELEMENT_TIMEOUT;
-		return helpers.listenFor(driver, 'player_element_exists', timeout).catch(function() {
+		return driver.waitForPlayerElement(timeout).catch(function() {
 			throw new Error('Player element is missing');
 		});
 	});
 }
 
-function promiseLoadPlayListen(driver, options) {
+/**
+ * @see {@link shouldBehaveLikeMusicSite}
+ * @return {Promise} Promise that will be resolved when the task has completed
+ */
+function promiseBehaveLikeMusicSite(driver, options) {
 	var opts = options || {};
 
 	return promiseLoadSite(driver, opts).then(function() {
