+++ tests/connectors/bbcradioplayer.js
@@ -1,9 +1,7 @@
 'use strict';
 
-/* global connectorSpec */
-
-module.exports = function(driver) {
-	connectorSpec.loadCheckPlayer(driver, {
+module.exports = function(driver, connectorSpec) {
+	connectorSpec.shouldContainPlayerElement(driver, {
 		url: 'http://www.bbc.co.uk/radio/player/bbc_6music'
 	});
 };
