+++ tests/connectors/youtube.js
@@ -1,9 +1,7 @@
 'use strict';
 
-/* global connectorSpec */
-
-module.exports = function(driver) {
-	connectorSpec.loadPlayListen(driver, {
+module.exports = function(driver, connectorSpec) {
+	connectorSpec.shouldBehaveLikeMusicSite(driver, {
 		url: 'https://www.youtube.com/watch?v=YqeW9_5kURI'
 	});
 };
